take error class into account when generating fingerprint
previously, this reporter was grouping together errors only using the
stack trace. this was insufficient for our programs because errors can
have similar stack traces when lacking pkg/errors backtraces.

the rollbar documentation mentions using the error class as part of the
fingerprint, so it makes sense to conform to that.